ctrlpt GENA: directly pass the device SID to the client code.
The lib used to generate a separate SID for the client side with a
one-to-one correspondance. The underlying idea was probably that several
client modules could share a device subscription, but this was never
actually implemented (would have needed a separate table), and we have a
single CP client anyway.