fix warnings under MinGW
The first problem is caused by the attribute not being matched properly
(gnu_printf vs printf). Simplest solution is to use the internal macro.

Second and third are wrong formats under Windows. For the third,
instead of using snprintf to write to a char buffer, it's simpler to
just use std::to_string.

Signed-off-by: Rosen Penev <rosenp@gmail.com>