Get rid of struct upnp_timeout
This was actually just used to share a 'handle' member. It was used only
in two places, and was not even used everywhere we use the TimerThread
event scheduling (ssdp did not use it). Inserting the appropriate data members
directly in the JobWorkers yields a lot of simplification and less memory allocations.